requirements.txt
boto==2.49.0
boto3==1.9.225
botocore==1.12.225
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.0.3
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
MarkupSafe==1.1.1
psycopg2==2.7.6.1
python-dateutil==2.8.0
python-dotenv==0.10.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
smart-open==1.8.4
urllib3==1.25.3
Werkzeug==0.15.4
wincertstore==0.2
